rename file: report-generator-v2 -> report-generator-impl
