+++ src/DetailsView/reports/report-generator-impl.ts
