+++ src/tests/unit/tests/DetailsView/reports/report-generator-provider.test.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 import { IMock, Mock } from 'typemoq';
 import { AssessmentReportHtmlGenerator } from '../../../../../DetailsView/reports/assessment-report-html-generator';
+import { ReportGeneratorImpl } from '../../../../../DetailsView/reports/report-generator-impl';
 import { createReportGeneratorProvider, ReportGeneratorProvider } from '../../../../../DetailsView/reports/report-generator-provider';
-import { ReportGeneratorImpl } from '../../../../../DetailsView/reports/report-generator-v2';
 import { ReportHtmlGenerator } from '../../../../../DetailsView/reports/report-html-generator';
 import { ReportNameGenerator } from '../../../../../DetailsView/reports/report-name-generator';
